Pull everything out to a dedicated 'RttSampler' class
