+++ dialogue-core/src/main/java/com/palantir/dialogue/core/RttSampler.java
@@ -1,0 +1,229 @@
+/*
+ * (c) Copyright 2020 Palantir Technologies Inc. All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.palantir.dialogue.core;
+
+import com.github.benmanes.caffeine.cache.Ticker;
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.collect.ImmutableList;
+import com.google.common.util.concurrent.FutureCallback;
+import com.google.common.util.concurrent.Futures;
+import com.google.common.util.concurrent.ListenableFuture;
+import com.google.common.util.concurrent.MoreExecutors;
+import com.palantir.dialogue.Endpoint;
+import com.palantir.dialogue.HttpMethod;
+import com.palantir.dialogue.Request;
+import com.palantir.dialogue.UrlBuilder;
+import com.palantir.logsafe.Preconditions;
+import com.palantir.logsafe.SafeArg;
+import java.io.Closeable;
+import java.time.Duration;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Map;
+import java.util.Optional;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicLong;
+import java.util.function.LongSupplier;
+import java.util.stream.Collectors;
+import java.util.stream.IntStream;
+import javax.annotation.concurrent.ThreadSafe;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+final class RttSampler {
+    private static final Logger log = LoggerFactory.getLogger(RttSampler.class);
+
+    private final ImmutableList<LimitedChannel> channels;
+    private final RttMeasurement[] rtts;
+    private final RttMeasurementRateLimiter rateLimiter;
+    private final Ticker clock;
+
+    RttSampler(ImmutableList<LimitedChannel> channels, Ticker clock) {
+        this.channels = channels;
+        this.rateLimiter = new RttMeasurementRateLimiter(clock);
+        this.clock = clock;
+        this.rtts = IntStream.range(0, channels.size())
+                .mapToObj(i -> new RttMeasurement())
+                .toArray(RttMeasurement[]::new);
+    }
+
+    /**
+     * The {@link LongSupplier} returns the RTT in nanoseconds, or {@link java.lang.Long#MAX_VALUE} to signify unknown.
+     */
+    LongSupplier get(int index) {
+        return rtts[index];
+    }
+
+    /**
+     * Non-blocking.
+     */
+    void maybeSampleRtts() {
+        Optional<RttMeasurementPermit> maybePermit = rateLimiter.acquireNonBlocking();
+        if (!maybePermit.isPresent()) {
+            return;
+        }
+
+        List<ListenableFuture<Long>> futures = IntStream.range(0, channels.size())
+                .mapToObj(i -> {
+                    long before = clock.read();
+                    return channels.get(i)
+                            .maybeExecute(
+                                    RttEndpoint.INSTANCE, Request.builder().build())
+                            .map(future -> Futures.transform(
+                                    future,
+                                    _response -> {
+                                        long durationNanos = clock.read() - before;
+                                        rtts[i].addMeasurement(durationNanos);
+                                        return durationNanos;
+                                    },
+                                    MoreExecutors.directExecutor()))
+                            .orElseGet(() -> Futures.immediateFuture(Long.MAX_VALUE));
+                })
+                .collect(ImmutableList.toImmutableList());
+
+        // the RttSamplePermit ensures that if a server black-holes one of our OPTIONS requests, we don't kick off
+        // more and more and more requests and eventually exhaust a threadpool
+        DialogueFutures.addDirectCallback(Futures.allAsList(futures), new FutureCallback<List<Long>>() {
+            @Override
+            public void onSuccess(List<Long> result) {
+                maybePermit.get().close();
+
+                if (log.isDebugEnabled()) {
+                    List<Long> millis =
+                            result.stream().map(TimeUnit.NANOSECONDS::toMillis).collect(Collectors.toList());
+                    long[] best = Arrays.stream(rtts)
+                            .mapToLong(rtt -> rtt.getAsLong())
+                            .toArray();
+                    log.debug(
+                            "RTTs {} {} {}",
+                            SafeArg.of("nanos", result),
+                            SafeArg.of("millis", millis),
+                            SafeArg.of("best", Arrays.toString(best)));
+                }
+            }
+
+            @Override
+            public void onFailure(Throwable throwable) {
+                maybePermit.get().close();
+                log.info("Failed to sample RTT for channels", throwable);
+            }
+        });
+    }
+
+    @VisibleForTesting
+    enum RttEndpoint implements Endpoint {
+        INSTANCE;
+
+        @Override
+        public void renderPath(Map<String, String> _params, UrlBuilder _url) {}
+
+        @Override
+        public HttpMethod httpMethod() {
+            return HttpMethod.OPTIONS;
+        }
+
+        @Override
+        public String serviceName() {
+            return "Balanced";
+        }
+
+        @Override
+        public String endpointName() {
+            return "rtt";
+        }
+
+        @Override
+        public String version() {
+            return "0.0.0";
+        }
+    }
+
+    /**
+     * Always returns the *minimum* value from the last few samples, so that we exclude slow calls that might include
+     * TLS handshakes.
+     */
+    @VisibleForTesting
+    @ThreadSafe
+    static final class RttMeasurement implements LongSupplier {
+        private static final int NUM_MEASUREMENTS = 5;
+
+        private final long[] samples;
+        private volatile long bestRttNanos = Long.MAX_VALUE;
+
+        RttMeasurement() {
+            samples = new long[NUM_MEASUREMENTS];
+            Arrays.fill(samples, Long.MAX_VALUE);
+        }
+
+        @Override
+        public long getAsLong() {
+            return bestRttNanos;
+        }
+
+        synchronized void addMeasurement(long newMeasurement) {
+            Preconditions.checkArgument(newMeasurement > 0, "Must be greater than zero");
+            Preconditions.checkArgument(newMeasurement < Long.MAX_VALUE, "Must be less than MAX_VALUE");
+
+            if (samples[0] == Long.MAX_VALUE) {
+                Arrays.fill(samples, newMeasurement);
+                bestRttNanos = newMeasurement;
+            } else {
+                System.arraycopy(samples, 1, samples, 0, NUM_MEASUREMENTS - 1);
+                samples[NUM_MEASUREMENTS - 1] = newMeasurement;
+                bestRttNanos = Arrays.stream(samples).min().getAsLong();
+            }
+        }
+
+        @Override
+        public String toString() {
+            return "RttMeasurement{" + "bestRttNanos=" + bestRttNanos + ", samples=" + Arrays.toString(samples) + '}';
+        }
+    }
+
+    private static final class RttMeasurementRateLimiter {
+        private static final long BETWEEN_SAMPLES = Duration.ofSeconds(1).toNanos();
+
+        private final Ticker clock;
+        private final AtomicLong lastMeasured = new AtomicLong(0);
+        private final AtomicBoolean currentlySampling = new AtomicBoolean(false);
+
+        private RttMeasurementRateLimiter(Ticker clock) {
+            this.clock = clock;
+        }
+
+        Optional<RttMeasurementPermit> acquireNonBlocking() {
+            long last = lastMeasured.get();
+            if (last + BETWEEN_SAMPLES > clock.read()) {
+                return Optional.empty();
+            }
+
+            if (currentlySampling.compareAndSet(false, true)) {
+                lastMeasured.set(clock.read());
+                return Optional.of(() -> currentlySampling.compareAndSet(true, false));
+            } else {
+                log.warn("Wanted to sample RTTs but an existing sample was still in progress");
+                return Optional.empty();
+            }
+        }
+    }
+
+    private interface RttMeasurementPermit extends Closeable {
+        @Override
+        void close();
+    }
+}
